server/requirements.txt
mcp>=1.2.0
httpx>=0.24.0
python-dotenv
diskcache
jsonschema